prototype beat detection
Computes spectral flux for the sample data, computes a threshold
and uses it to isolate peaks, which are our beats.
See https://www.badlogicgames.com/wordpress/?p=187 for more information
More work needs to be done in verifying and tuning the algorithim.
In NonUnity/graphFlux.html, began visiulizing the output peaks.