Initial support for BeatMap generation on Android
Use a java plugin to do this, CustomUnityPlayerActivity,
which replaces the default UnityPlayerActivity.
This file then opens the android native file picker and copies the
specified content uri's content to a file in /Songs/.
Changed BeatMap to load from a song in /Songs/.
Added buttons to go from main menu to BGAScene and vice versa.
BGAScene still WIP and will eventually moved to main menu.